manifest: add a container source to OSTreeDeployment
Support a container source for OSTreeDeployment as an alternative to the
ostree commit source.
Currently, if a container is defined, it fails with an error.

The commit source and the container source are now both pointers since
they can be unset, but (will) we need to make sure at least one is set
when serializing.