manifest: label {,u}mount as `install_exec_t` in buildroot
When running osbuild inside a container created via the new
`osbuild-deploy-container` [1] the selinux setup is slightly
different from a non container setup and this needs to be
accounted for:

Inside the container all files in the overlayfs have the
label `system_u:object_r:container_files_t`. Because the
container is using an overlayfs individual files like
`/usr/bin/osbuild` cannot be (re)labled `osbuild_exec_t`
as on the normal system. But osbuild needs CAP_MAC_ADMIN
to label it's buildroot correctly.

This is why osbuild-deploy-container creates a tmpdir and
copies/relabels osbuild there as `install_exec_t` to ensure
that osbuild runs with a selinux domain powerful enough to
have CAP_MAC_ADMIN (this is required to set unknown selinux
labels).

The normal `osbuild_t`->`install_t` transition does not work
because `/` inside the container is mounted `nosuid` which
will prevent also transitionsto more "powerful" selinux
contexts like osbuild_t->install_t.

Yet still selinux denial errors are in the logs when the
buildroot uses the system selinux labeling and use the
labels `mount_t` for `{,u}mount` inside the build root.
The `install_t` context is not allowed to transition to
`mount_t` however it is powerful enough to perform the
mount operations so the error has no impact on the operation.

This commit fixes this error entirely by labeling `{,u}mount`
inside the buildroot also with `install_exec_t` to ensure
that no transition is required.

[1] https://github.com/osbuild/osbuild-deploy-container/