otk-gen-partition-table: add UUID and fstab props to input partition
Add UUID and fstab-related properties to the input partition object.

The UUID lets users define partition UUIDs on their base partition
tables instead of letting the rng make one for them.

The fstab-related properties are used when generating the fstab file
with the org.osbuild.fstab stage.  The json tags (fs_) are the names of
the fields used in the fstab(5) manpage.