disk: add defaultType to NewCustomPartitionTable()
Add a new argument, defaultType, to the NewCustomPartitionTable()
function.  The purpose of the type is to determine which filesystem will
be used for root and boot when they are added automatically.

It will also be used to set the filesystem for any mountpoint that has
no type defined.

The default type should be determined by the distro definition.