disk: add PartitionTableType to CustomPartitionTableOptions
Add an option for the partition table type to the
CustomPartitionTableOptions.  The value must be one of "gpt", "dos", or
"".  An empty string (or nil options) defaults to "gpt".

The tests have been updated to explicitly specify "dos".  The relayout()
method, which is affected by this, checks whether `pt.Type == "gpt"`,
which means it was defaulting to "dos" when deciding to add a footer.