blueprint: validate partitioning customizations
Add a Validate() method to PartitioningCustomization that returns an
error if:
1. Btrfs and LVM are both defined.
2. Any mountpoint is invalid (empty string, relative path, or not
   "clean").
3. Any mountpoint is defined more than once in the whole structure.
4. There is more than one LVM Volume Group or more than one btrfs
   volume.
5. Any LVM Logical Volume name is used more than once in the same
   VolumeGroup.
6. Any Btrfs subvolume has an empty name.
7. Any Btrfs subvolume name is used more than once in the same Btrfs
   volume.

This function validates the customization itself in the absence of any
distro or image-type specific policies.

We may relax rule 3 in the future, at least in images if not in
osbuild-composer, but for now, let's disallow it while keeping the
customization structure flexible.

val