disk: add RequiredSizes to CustomPartitionTableOptions
Introduce the concept of required sizes from the existing
NewPartitionTable() function into the NewCustomPartitionTable().

RequiredSizes is necessary for setting a "sane default" (minimum) size
for certain directories.  By default, it forces the partition table to
reserve 1 GiB for / and 2 GiB for /usr, which in most cases (when there
is no separate /usr filesystem) means that the root filesystem is at
least 3 GiB.

Without this mechanism, automatically created root filesystems can end
up being too small to fit the base OS packages or to have a usable
system in the long run.  It's also needed when automatically creating
root partitions or LVM Logical Volumes.  Without a required minimum, the
root partition will only be the size of the leftover space on the image
(the base image size minus the sum of the rest of the partitions, which
can sometimes end up being 0) or LVM VolumeGroup.

Update tests.