disk: generate volume group names when unspecified
When a volume group's name is not specified in the customizations, the
custom partition table creator assigned the names vg00, vg01, etc.

A new test is added with multiple volume groups to test this.  The
scenario of the test is not supported by customizations, the blueprint
package does not allow multiple volume groups, but the function can
handle it so it's good to have it covered.