repositories: support "alias" feature in repo json
This is a hack to workaround the lack of symlinks in golang embedd.
Traditionally we used symlinks to alias distro names (like centos-10
and centos-stream-10 being the same). But we cannot use symlinks when
using golangs embedd feature so we implement this now via an "alias"
in the json itself.

The alternative is to rework the existing distributions json and
make it more structured, e.g.
```json
{
 "distro_name": "centos-10",
 "distro_aliases": ["centos-stream-10", "funky-10"],
 "repos": { ...existing structre...}
...
```
which of course requires a bit compatibility work.