disk: extract `getDevicesNoRename()` and tweak `getDevices()`
We currently do a lot of:
```
lastDevice := stageDevices[lastName]
delete(stageDevices, lastName)
stageDevices["device"] = lastDevice
```
which this commit consolidates into the `getDevices()` helper.
The only place that needs the "old" behavior is
`GenMountsDevicesFromPT` and here we can use the new
`getDevicesNoRename()` helper.

Note that this is still pretty messy and needs further refactor.