many: pass `dnfjson.DepsolveResult` into `manifest.Serialize`
This is a quick draft to show what it make manifest.Serialize()
less "floppy", i.e. to make it
a) type-safe so that only depsolved pkgs/repos can be passed
b) ensure pkg/repos that got depsolved together are grouped together

Note that this is reusing `dnfjson.DepsolveResult` for simplicity
to how the idea its an open question if we should actually use
it. The upside is that it gives us e.g. the sbom everyhwere which
might be nice. The downside is that its not tailored and a bit
long (dnfjson.Result might be enough for this already).