cmd: make `cmd/build` use `manifestgen`
This commit makes the `cmd/build` command use the manifestgen module
to remove some code duplication.

It does not move to use the actual `manifestgen.Generator` because
the way the repositories are used is not supported by manifestgen
right now: `build` supports loading a single repository file but
manifestgen expects a reporegistry so that it can map "distroname"
to a "repository".

We could add support for loading a repo json directly in
`manifestgen.Generator` but I do feel uneasy about it because it
seems a good property that there is some connection between the
distroname and the repo in the general case.