imagefilter: make supportedFilters less "loose"
This commit makes the supported filters part of a map so that
we can never get them out of sync. The downside is that this
way we also no longer control the order of the filters (because
maps are not sorted). This is an important property though because
most people will care about "distro" much more than about "bootmode".