osbuild: use a `bufio.Reader` instead of a `Scanner`
We ran into the 64kb buffer limit with the bufio.Scaner recently.
The fact that the scanner can also not recover from that is a bit
daunting. Because we really just need something that gives us the
next `\n` we switch to a `bufi.Reader` that does not have this
limitation (it will happy OOM on endless input though, you can't
win).