many: introduce new `inputs.RepoContainerConfig`
This commit changes the signature of `image.InstantiateManifest`
so that it no longer takes a []RepoConfig as input to instanciate
the image but instead a new struct that contains both a []RepoConfig
and a buildroot container reference.

With that we can let each distro also create an (optional) buildroot
reference.