packagesets: add basic imagetype yaml to merge pkgSets
This commit adds a basic imagetype yaml to support merging
of package sets  This is a useful feature so that we can e.g.
avoid duplicating the `cloud_base` or `anaconda` package sets.

It works by adding a new toplevel "image_types" in the
package_sets.yaml and then using anchors/refs to refer to them
in the image_type yaml.

It allows us to write:
```yaml
package_sets:
  cloud_base_pkgset: &cloud_base_pkgset
    include: [foo]
    exclude: [bar]
  qcow2_pkgset: &qcow2_pkgset
    include: ["qemu-guest-agent"]

image_types:
 qcow2:
  package_sets:
   - *cloud_base_pkgset
   - *qcow2_pkgset
```
The downside of this approach is that:
1. it requires the anchor to be there before the ref, which can
   be confusing (we could solve this later via includes)
2. if the anchor is forgotten the error becomes very obscure

The upside is that it is very symetrical.