many: add `UseBootstrapContainer` option to manifestgen
This commit adds a new option `UseBootstrapContainer` to
manifestgen that is then used during manifest creation to
add a `DistroBootstrapRef` to the manifest. When this is
set the image functions need to honor it and create an
appropriate buildroot that contains a bootstrap pipeline.
This work is part of the next commit.