distro: add defs.ImageTypes() to load all image types
This commit adds support to load the full image types
definition directly from YAML. Note that this does not
(yet) implement the `distro.ImageType` interface, it
just provides the ability to load the base information
about the image type and then the imageTypeYAML needs
to be converted into the "native" fedora/rhel image
type. This will change in a subsequent refactor and
there will be only a single (yaml based) image type
implementation.