distro: add `distros.yaml` that define the distros in pure YAML
This commit moves the distro definitions out of code into pure
YAML. This adds a bit of redundancy because there is no code
anymore to help with auto-creating the distributions if it
finds the right repositories. But the upside is that the hirarchy
is more well defined now and to add a new distro or derrive a
distro it is enough to just add a new list item into `distros.yaml`
and it will get picked up.

The redundancy is also not too bad when YAML anchors are used,
e.g.:
```yaml
  - &rhel_9
    name: rhel-9.7
    os_version: "9.7"
    product: "Red Hat Enterprise Linux"
    release_version: "9"
    module_platform_id: "platform:el9"
    vendor: "redhat"
    ostree_ref_tmpl: "rhel/9/%s/edge"
    defs: "rhel-9"
  - <<: *rhel_9
    name: rhel-9.6
    os_version: "9.6"
  - <<: *rhel_9
    name: rhel-9.5
    os_version: "9.5"
...
```

We should probably also move the repositories json into this file so
that its all in a single file.

But it may become a bit unwieldy as the gpg keys add quite a bit of lines.
OTOH adding GPG via YAML would allow them to be nicer than JSON, e.g.:
```yaml
distros:
  - name: rhel-9.7
    repos:
      aarch64:
        - name: baseos
          baseurl: "https://cdn.redhat.com/content/dist/rhel9/9.7/aarch64/baseos/os"
          rhsm: true
          check_gpg: true
          gpgkey: |
            -----BEGIN PGP PUBLIC KEY BLOCK-----
            Version: GnuPG v2.0.22 (GNU/Linux)
            mQINBErgSTsBEACh2A4b0O9t+vzC9VrVtL1AKvUWi9OPCjkvR7Xd8DtJxeeMZ5eF
            ...
            -----END PGP PUBLIC KEY BLOCK-----
...
```