distro: note about moving DistroImageConfig into distros.yaml
Add a note about moving the default distro image config into
distros.yaml or keep it in the image types file.

Keeping it in the image types file has the advantage that
things are "in-sync" naturally, i.e. when having a
`rhel-10/images.yaml` and `distros.yaml` like this:
```yaml
distros:
 - name: "centos-10"
   defs_dir: ./rhel-10
 - name: "rhel-10.0"
   defs_dir: ./rhel-10
```
there is no need to duplicate the default distro config. OTOH
it feels more "natural" to have the DefaultImageConfig as
part of the above.