image: add conditions as strings via starlark
The rhel-9 azure image type needs logical operations for its
image config. It checks for example that the arch is x86_64
and the rhel version >= 9.6. Expressing this via pure YAML
becomes rather cumbersome, so after some experiments this
PR seems to be a really nice way forward.

The idea is to use `starlark` which is a dialect of python
optimized for configuration languages and used in e.g.
bezel. With that we can simplify our conditionals to be
python expressions and they now look like:
```yaml
image_types:
  "workstation-live-installer":
    ...
    image_config:
      locale: "en_US.UTF-8"
      iso_rootfs_type: "squashfs"
      conditions:
        "f40 and below uses ext4 based iso rootfs":
          when: version_less_than("41")
          merge:
            iso_rootfs_type: "squashfs-ext4"
```
The following keys are supported inside the `when`:
- distro_name
- distro_version
- distro_major
- distro_minor
- arch
And the following helper functions:
- version_less_than()
- version_greater_or_equal()
With that we can express all our intend nicely and we can
also express what a "when" is doing, i.e. for some (like
package sets) it will append, for others (image-types) it
will merge and for installer-config it will override. All
of this is now easier to read via:
```yaml
when: <expr>
override: ...
``