gen-manifests: add `-bootc-ref` switch to enable bootc manifests
Add a `--bootc-ref` so that we can generate manifests for bootc
container based images.

This allows us to write:
```console
$ gen-manifests -arches x86_64 -types qcow2 -bootc-refs quay.io/centos-bootc/centos-bootc:stream9 -metadata=false  -distros ignore -config bp.json
```
which then can be used to diff the manifests from images with
the manifests that bib produces.

It also is a first step to have manifest checksums for bootc
manifests.