distro: bring per-distro checkOptions back
We actually need to check options based on the "distro-like"
distro major version. Between rhel8 and rhel9 it seems some
customizations are enabled that did not work before so
until we have a more declarative way (and more tests) for
the options checking we move a step "back" and bring back
a lot of duplicated code from the original options checks.