distro: add container image for leap-16.0
This is an experiment to see how hard it is to add a container
image in opensuse leap. It is actually hard, we have a bunch
of things easy enough to fix in images (this commit includes
them just to highlight what would be needed).

But the blocker is that osbuild/bwrap do not work, the way we
do it is that we setup an empty /etc when we run the buildroot.
But opensues has a libz-ng-compat rpm that requires a custom
ld.so.conf:
```
bash-5.2# cat /etc/ld.so.conf && echo ------------ && cat /etc/ld.so.conf.d/zlib-ng-compat.conf
/usr/local/lib64
/usr/local/lib
include /etc/ld.so.conf.d/*.conf
------------
/usr/lib64/zlib-ng-compat
```
which is currently not supported osbuild. We could workaround
it via a custom runner in osbuild that replicates the creation
of the suse ld.so.conf and then bind mount the ld.so.conf.d
but its quite involved.

osbuild will need something like:
```python
diff --git a/osbuild/buildroot.py b/osbuild/buildroot.py
index 02b1b9f2..a2f52110 100644
--- a/osbuild/buildroot.py
+++ b/osbuild/buildroot.py
@@ -196,7 +196,8 @@ class BuildRoot(contextlib.AbstractContextManager):
         mounts = []

         # Import directories from the caller-provided root.
-        imports = ["usr"]
+        # XXX: opensuse needs /etc/ld.so.conf.d
+        imports = ["usr", "etc/ld.so.conf.d"]
         if self.mount_boot:
             imports.insert(0, "boot")

diff --git a/runners/org.osbuild.leap b/runners/org.osbuild.leap
new file mode 100755
index 00000000..f57290d6
--- /dev/null
+++ b/runners/org.osbuild.leap
@@ -0,0 +1,16 @@
+#!/usr/bin/python3
+
+import subprocess
+import sys
+
+from osbuild import api
+
+
+if __name__ == "__main__":
+    with api.exception_handler():
+        with open("/etc/ld.so.conf", "w") as fp:
+           fp.write("include /etc/ld.so.conf.d/*.conf\n")
+        subprocess.run(["ldconfig"], check=True)
+
+        r = subprocess.run(sys.argv[1:], check=False)
+        sys.exit(r.returncode)
```