manifest: rework Exports via ExportingPipeline
Rework GetExports() to just use the static information,
i.e. just collect all `ExportingPipelines` of the manifest.
This removes some code. It is arguably less correct than
before because before we only collected the export that
actually ran during manifest serialization. It is arguably
more correct because before it was possible to have "dead"
pipelines that could potentially export but are never noticed.