many: drop the workload.Workload type entirely
Workloads are no longer used in the images library and there
is no appetite to use them. So this commit removes the last
use of them: the workload.Custom{} that was used to pass
image type configuration to the images functions.

The workload.Custom is replaces with the regular
manifest.OSCustomizations that already contains (almost)
anything that is needed. So instead of having two concepts
to configure the image we consolidate it to one.

The only thing that had to be added to the OSCustomizations
is the `Modules`.