distro,image: drop `imgTypeCustomizations`
In https://github.com/osbuild/images/pull/1770 we did the
first step to remove the workload and replaced them with
the (badly named) `imgTypeCustomizations`. This commit
now removes this too and folds it into the existing
OSCustomizations for clarity and to have one less thing
to worry about.

The payload repositories are now passed to the image
functions and get added to the OSCustomizations there
as a new `PayloadRepos`. They will be used when resolving
blueprint packages.