test: add vendored `vm.py`
This commit vendors the bootc-image-builder vm abscration:
https://raw.githubusercontent.com/osbuild/bootc-image-builder/refs/heads/main/test/vm.py

so that we can use it for local boot tests. Note that this is
similar to test.thing that is used in
https://github.com/osbuild/images/pull/1982

But it uses a very traditional approach, i.e. require network and
no vsocket. The issue with test.thing is that we need to support
rhel-{8,9,10} and fedora and the policyfill approach is quite a
bit of extra work for us but we don't get any benefits from the
VirtualMachineAPI.md as none of our images supports it yet.

We should probably look into making test.thing support legycy
(non vsocket) mode?