distro,manifestgen: support custom depsolvers for distros
The bootc based distros need a custom depsolver because
they need to use the dnf from the container to ensure
that all the plugins/custom config is used.

So manifestgen now checks if the distro defines its own
way to create a depsovler and if so passes it into the
manifestgen.Depsolver.

With that bootc ISOs (that require rpms) are compatible with
custom depsolve helpers and we can use a custom depsolver
to catpure the mTLS keys for the rpms that are required
by bootc.